Edit configs, move functions into utils dir, add parcel
