+++ packages/core-v2-sdk/.env.example
@@ -1,3 +1,5 @@
 ETHERSCAN_API_KEY=ABC123ABC123ABC123ABC123ABC123ABC1
-ROPSTEN_URL=https://eth-ropsten.alchemyapi.io/v2/<YOUR ALCHEMY KEY>
-PRIVATE_KEY=0xabc123abc123abc123abc123abc123abc123abc123abc123abc123abc123abc1
+MAINNET_URI=https://eth-mainnet.alchemyapi.io/v2/<YOUR ALCHEMY KEY>
+GOERLI_URI=https://eth-goerli.alchemyapi.io/v2/<YOUR ALCHEMY KEY>
+# "mainnet" | "mainnet-fork" | "goerli" | "testnet"
+CHAIN_NAME=mainnet
