+++ packages/core-v2-sdk/.eslintignore
@@ -2,3 +2,4 @@ node_modules
 artifacts
 cache
 coverage
+dist
