+++ packages/core-v2-sdk/.prettierignore
@@ -2,4 +2,4 @@ node_modules
 artifacts
 cache
 coverage*
-gasReporterOutput.json
+dist
