--- packages/core-v2-sdk/.solhintignore
@@ -1,1 +1,0 @@
-node_modules
