+++ packages/core-v2-sdk/src/utils/calcSwapCongergentCurvePool/calculateTradePrincipalTokens.ts
